fix: refactored singleFriend redux file
